Make Website CMS use DefaultTemplate.html (#29)
Remove bypassing middleware if the URL starts with /website/cms
Remove including a Self.GET to the user area, because DefaultTemplate will trigger it through mapping.
The above change allowed to remove /Website/user request handler and URI mapping to it.
WrapperPage now stores a reference to the previous PublicViewModel, so CMS can find its pages that registered a commit hook.